Move documents to new folder in schemas
