+++ studio/schemas/documents/project.ts
